@@ -1,21 +1,22 @@
 module github.com/microfast-ch/rea
 
-go 1.18
+go 1.21
 
 require (
 	github.com/Shopify/go-lua v0.0.0-20220120202609-9ab779377807
-	github.com/davecgh/go-spew v1.1.1
+	github.com/Shopify/goluago v0.0.0-20210621135517-fe0528d0b204
+	github.com/djboris9/xmltree v0.0.0-20220419153310-7195f31abe9c
+	github.com/google/go-cmp v0.5.7
+	github.com/spf13/cobra v1.4.0
+	github.com/stretchr/testify v1.7.1
+	golang.org/x/exp v0.0.0-20220328175248-053ad81199eb
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c
 )
 
 require (
-	github.com/Shopify/goluago v0.0.0-20210621135517-fe0528d0b204 // indirect
-	github.com/djboris9/xmltree v0.0.0-20220419153310-7195f31abe9c // indirect
-	github.com/google/go-cmp v0.5.7 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/inconshreveable/mousetrap v1.0.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/spf13/cobra v1.4.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
-	github.com/stretchr/testify v1.7.1 // indirect
-	golang.org/x/exp v0.0.0-20220328175248-053ad81199eb // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
 )
